typo fixes

+++ lab08/lab08.cpp
@@ -12,7 +12,7 @@
 
 double runge_eps = 0;
 double adams_eps = 0;
-int n = 20;
+size_t n = 20;
 double a = -1.5;
 double b = 0.5;
 
@@ -38,7 +38,7 @@ double test_ans(double x) {
     return 2 * exp(x);
 }
 
-vector<double> rungeKuttaSolution(double (*fxu)(double x, double y), double a, double b, double n) {
+vector<double> rungeKuttaSolution(double (*fxu)(double x, double y), double a, double b, size_t n) {
     vector<double> res;
     double h = (b - a) / n;
     res.reserve(n + 1);
@@ -66,7 +66,7 @@ vector<double> rungeKuttaSolution(double (*fxu)(double x, double y), double a, d
     return res;
 }
 
-vector<double> adamsSolution(double (*fxu)(double x, double y), double a, double b, double n, double y0,
+vector<double> adamsSolution(double (*fxu)(double x, double y), double a, double b, size_t n, double y0,
                              double y1, double y2, double y3) {
     vector<double> res;
     double h = (b - a) / n;
@@ -102,7 +102,7 @@ vector<double> adamsSolution(double (*fxu)(double x, double y), double a, double
     return res;
 }
 
-vector<double> realSolution(double (*fx)(double x), double a, double b, double n) {
+vector<double> realSolution(double (*fx)(double x), double a, double b, size_t n) {
     vector<double> res;
     double h = (b - a) / n;
     res.reserve(n + 1);
@@ -155,33 +155,36 @@ vector<double> real() {
 }
 
 
-vector<pair<double,double>> rungeKuttaSystemSolution(double (*fxu1)(double x, double y1, double y2),double (*fxu2)(double x, double y1, double y2), double a, double b, double n) {
-    vector<pair<double,double>> res;
+vector<pair<double, double>> rungeKuttaSystemSolution(double (*fxu1)(double x, double y1, double y2),
+                                                      double (*fxu2)(double x, double y1, double y2), double a,
+                                                      double b, size_t n) {
+    vector<pair<double, double>> res;
     double h = (b - a) / n;
     res.reserve(n + 1);
 
-	res.push_back(make_pair(fxu1(a, 1,1),fxu2(a, 1,1))); // first elem;
+    res.push_back(make_pair(fxu1(a, 1, 1), fxu2(a, 1, 1))); // first elem;
 
     for (int i = 0; i < n; i++) {
-		double yn = res.back().first;
-		double yn2 = res.back().second;
+        double yn = res.back().first;
+        double yn2 = res.back().second;
         double xn = a + i * h;
 
         double k11 = fxu1(xn, yn, yn2);
-		double k12 = fxu2(xn, yn, yn2);
+        double k12 = fxu2(xn, yn, yn2);
 
         double k21 = fxu1(xn + h / 2, yn + (h * k11) / 2, yn2 + (h * k12) / 2);
-		double k22 = fxu2(xn + h / 2, yn + (h * k11) / 2, yn2 + (h * k12) / 2);
+        double k22 = fxu2(xn + h / 2, yn + (h * k11) / 2, yn2 + (h * k12) / 2);
 
         double k31 = fxu1(xn + h / 2, yn + (h * k21) / 2, yn2 + (h * k22) / 2);
-		double k32 = fxu2(xn + h / 2, yn + (h * k21) / 2, yn2 + (h * k22) / 2);
+        double k32 = fxu2(xn + h / 2, yn + (h * k21) / 2, yn2 + (h * k22) / 2);
 
         double k41 = fxu1(xn + h, yn + (h * k31), yn2 + (h * k32));
-		double k42 = fxu2(xn + h, yn + (h * k31), yn2 + (h * k32));
+        double k42 = fxu2(xn + h, yn + (h * k31), yn2 + (h * k32));
 
-        res.push_back(make_pair(yn + (k11 + 2 * k21 + 2 * k31 + k41) * h / 6 , yn2 + (k12 + 2 * k22 + 2 * k32 + k42) * h / 6));
-		
-	
+        res.push_back(
+                make_pair(yn + (k11 + 2 * k21 + 2 * k31 + k41) * h / 6, yn2 + (k12 + 2 * k22 + 2 * k32 + k42) * h / 6));
+
+
 //        double k1 = fxu(xn, yn);
 //        double k2 = fxu(xn + h / 4, yn + (h * k1) / 4);
 //        double k3 = fxu(xn + h / 2, yn + (h * k2) / 2);
@@ -193,23 +196,25 @@ vector<pair<double,double>> rungeKuttaSystemSolution(double (*fxu1)(double x, do
 }
 
 
-vector<pair<double,double>> adamsSystemSolution(double (*fxu1)(double x, double y1, double y2),double (*fxu2)(double x, double y1, double y2), double a, double b, double n,vector<pair<double,double>> kutta) {
-    vector<pair<double,double>> res;
+vector<pair<double, double>> adamsSystemSolution(double (*fxu1)(double x, double y1, double y2),
+                                                 double (*fxu2)(double x, double y1, double y2), double a, double b,
+                                                 size_t n, vector<pair<double, double>> kutta) {
+    vector<pair<double, double>> res;
     double h = (b - a) / n;
     res.reserve(n + 1);
 
-	res.push_back(kutta[0]); // first elem;
-	res.push_back(kutta[1]);
-	res.push_back(kutta[2]);
-	res.push_back(kutta[3]);
+    res.push_back(kutta[0]); // first elem;
+    res.push_back(kutta[1]);
+    res.push_back(kutta[2]);
+    res.push_back(kutta[3]);
 
     for (int i = 3; i < n; i++) {
-		double yn = (res.end() - 1)->first;
+        double yn = (res.end() - 1)->first;
         double yn_1 = (res.end() - 2)->first;
         double yn_2 = (res.end() - 3)->first;
         double yn_3 = (res.end() - 4)->first;
 
-		double _2yn = (res.end() - 1)->second;
+        double _2yn = (res.end() - 1)->second;
         double _2yn_1 = (res.end() - 2)->second;
         double _2yn_2 = (res.end() - 3)->second;
         double _2yn_3 = (res.end() - 4)->second;
@@ -220,10 +225,12 @@ vector<pair<double,double>> adamsSystemSolution(double (*fxu1)(double x, double 
         double xn_3 = a + (i - 3) * h;
 
         res.push_back(
-                make_pair(yn + h * (55 * fxu1(xn, yn, _2yn) - 59 * fxu1(xn_1, yn_1, _2yn_1) + 37 * fxu1(xn_2, yn_2,_2yn_2) - 9 * fxu1(xn_3, yn_3,_2yn_3)) / 24 ,
-				_2yn + h * (55 * fxu2(xn, yn, _2yn) - 59 * fxu2(xn_1, yn_1, _2yn_1) + 37 * fxu2(xn_2, yn_2,_2yn_2) - 9 * fxu2(xn_3, yn_3,_2yn_3)) / 24));
-		
-	
+                make_pair(yn + h * (55 * fxu1(xn, yn, _2yn) - 59 * fxu1(xn_1, yn_1, _2yn_1) +
+                                    37 * fxu1(xn_2, yn_2, _2yn_2) - 9 * fxu1(xn_3, yn_3, _2yn_3)) / 24,
+                          _2yn + h * (55 * fxu2(xn, yn, _2yn) - 59 * fxu2(xn_1, yn_1, _2yn_1) +
+                                      37 * fxu2(xn_2, yn_2, _2yn_2) - 9 * fxu2(xn_3, yn_3, _2yn_3)) / 24));
+
+
 //        double k1 = fxu(xn, yn);
 //        double k2 = fxu(xn + h / 4, yn + (h * k1) / 4);
 //        double k3 = fxu(xn + h / 2, yn + (h * k2) / 2);
@@ -236,14 +243,14 @@ vector<pair<double,double>> adamsSystemSolution(double (*fxu1)(double x, double 
 
 double test_fxu1(double x, double u, double u2) {
     if (x == 0) return 0.6;
-     return x*cos(u+u2);
+    return x * cos(u + u2);
 //    if (x == 0) return 0;
 //    return (x * x + 1) * u + cos(x) + (-1 - x * x) * sin(x);
 }
 
 double test_fxu2(double x, double u, double u2) {
     if (x == 0) return 2;
-    return sin(u-u2);
+    return sin(u - u2);
 //    if (x == 0) return 0;
 //    return (x * x + 1) * u + cos(x) + (-1 - x * x) * sin(x);
 }
@@ -264,28 +271,28 @@ int main() {
     cout << "runge eps " << runge_eps << endl;
     cout << "adams eps " << adams_eps << endl;
 
-	auto res = rungeKuttaSystemSolution(test_fxu1,test_fxu2,0,2,20);
-	for(int i = 0; i <=20;i++) {
-		cout << res[i].first << endl;
+    auto res = rungeKuttaSystemSolution(test_fxu1, test_fxu2, 0, 2, 20);
+    for (int i = 0; i <= 20; i++) {
+        cout << res[i].first << endl;
 
-	}
-	cout << endl << endl;
-		for(int i = 0; i <=20;i++) {
-		cout << res[i].second << endl;
+    }
+    cout << endl << endl;
+    for (int i = 0; i <= 20; i++) {
+        cout << res[i].second << endl;
 
-	}
+    }
 
-		cout << endl << endl;
+    cout << endl << endl;
 
-			auto res2 = adamsSystemSolution(test_fxu1,test_fxu2,0,2,20,res);
-	for(int i = 0; i <=20;i++) {
-		cout << res2[i].first << endl;
+    auto res2 = adamsSystemSolution(test_fxu1, test_fxu2, 0, 2, 20, res);
+    for (int i = 0; i <= 20; i++) {
+        cout << res2[i].first << endl;
 
-	}
-	cout << endl << endl;
-		for(int i = 0; i <=20;i++) {
-		cout << res2[i].second << endl;
+    }
+    cout << endl << endl;
+    for (int i = 0; i <= 20; i++) {
+        cout << res2[i].second << endl;
 
-	}
+    }
     return 0;
 }
